+++ src/main/java/com/owmax/model/Blanks.java
@@ -11,7 +11,7 @@ public class Blanks implements java.io.Serializable {
 
 	// Fields
 
-	private Integer id;
+	private Short id;
 	private Questions questions;
 	private String blankText;
 	private Set answerBlankses = new HashSet(0);
@@ -36,11 +36,11 @@ public class Blanks implements java.io.Serializable {
 
 	// Property accessors
 
-	public Integer getId() {
+	public Short getId() {
 		return this.id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Short id) {
 		this.id = id;
 	}
