+++ src/main/java/com/owmax/model/Questions.java
@@ -1,6 +1,5 @@
 package com.owmax.model;
 
-import javax.persistence.OrderBy;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -12,7 +11,7 @@ public class Questions implements java.io.Serializable {
 
 	// Fields
 
-	private Integer id;
+	private Short id;
 	private String title;
 	private Short type;
 	private Short maxSelection;
@@ -45,11 +44,11 @@ public class Questions implements java.io.Serializable {
 
 	// Property accessors
 
-	public Integer getId() {
+	public Short getId() {
 		return this.id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Short id) {
 		this.id = id;
 	}
